Set weak pull-up on RAS0
This change enables an internal (to the FPGA) weak pull-up on the RAS0 pin.
Once #11 is implemented in hardware, the RAS0 pin on the FPGA will either be
connected to the RAS0 pin on the trapdoor connector or be disconnected.
If RAS0 is disconnected then the weak pull-up makes sure the RAS0 is inactive.